third_party/WebKit/LayoutTests/inspector/console/console-export.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function log() {
    var nonUrl = "foo";
    var url = "www.chromium.org";
    var longNonUrl = "baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaar";
    var longUrl = "www.loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooongurl.com";
    console.log(nonUrl);
    console.log(url);
    console.log(longNonUrl);
    console.log(longUrl);
    console.log(url + " " + longUrl + " " + nonUrl + " " + longNonUrl + " " + url + " " + longUrl);
    console.trace("My important trace");
    console.error("My error");
}
log();

function test()
{
    InspectorTest.addResult("\nDumping messages");
    InspectorTest.dumpConsoleMessages();
    InspectorTest.addResult("\nDumping export strings");
    var consoleView = Console.ConsoleView.instance();
    consoleView._visibleViewMessages.forEach(message => InspectorTest.addResult(message.toExportString()));
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that exporting console messages produces proper output.
</p>
</body>
</html>
